\name{ATSList}
\alias{ATSList}
\title{A list of \code{\link{aTS}} objects}
\description{
  An ATSList is simply a \code{list} of \code{\link{aTS}}
  objects. However, there are a lot of useful functions
  that can be used together with such a list, namely
  \code{\link{setNamesATSList}},
  \code{\link{summaryTableATSList}},
  \code{\link{computePartitionsATSList}},
  \code{\link{groupATSListByPartitions}}, and
  \code{\link{embedATSList}}. Furthermore, for I/O, there
  are: \code{\link{loadDifToATSList}},
  \code{\link{loadDatDirToATSList}},
  \code{\link{loadDataDirToATSList}},
  \code{\link{loadATSDir}}, \code{\link{loadATSDir}},
  \code{\link{saveATSList}},
  \code{\link{savePlotsATSList}}, and
  \code{\link{saveAsDataPackageATSList}}.
}

