\name{cBindByRowNames}
\alias{cBindByRowNames}
\title{Column-bind together two tables, taking into account the row names}
\usage{
  cBindByRowNames(x, y)
}
\arguments{
  \item{x}{first table}

  \item{y}{second table}
}
\value{
  the two tables cbind-ed together
}
\description{
  The function is the pendant to
  \code{\link{rBindByColNames}}. See there for more
  detailed explanations.
}
\seealso{
  \code{\link{rBindByColNames}}
}

