\name{computePartitionsATSList}
\alias{computePartitionsATSList}
\title{Compute partitions for an \code{\link{ATSList}}}
\usage{
  computePartitionsATSList(atsList, ...)
}
\arguments{
  \item{atsList}{the \code{\link{ATSList}} to which to
  apply the function to}

  \item{...}{parameters passed to
  \code{\link{computePartitions}}}
}
\description{
  This method calls for every \code{\link{aTS}} object in
  the \code{\link{ATSList}} the function
  \code{\link{computePartitions}}.
}

