\name{denormalizeData}
\alias{denormalizeData}
\title{Revert data normalization.}
\usage{
  denormalizeData(x, normParams)
}
\arguments{
  \item{x}{input data}

  \item{normParams}{the parameters generated by an earlier
  call to \code{normalizeData} that will be used for
  reverting normalization}
}
\value{
  column-wise reverse-normalized input
}
\description{
  Column-wise normalization of the input matrix is
  reverted.
}

