\name{firstColToColname}
\alias{firstColToColname}
\title{Use the first column of a matrix to set its row names}
\usage{
  firstColToColname(x)
}
\arguments{
  \item{x}{a matrix}
}
\value{
  the matrix, but with the first column removed and set to
  rownames
}
\description{
  TODO: This function should be renamed to
  \code{firstColToRowNames}
}

