\name{getConfigDefault}
\alias{getConfigDefault}
\title{Get the default settings of parallelization config}
\usage{
  getConfigDefault()
}
\value{
  the default settings
}
\description{
  This function returns default settings for
  parallelization, to be used with
  \code{\link{initParallelMode}}. The default is no
  parallelization.
}
\seealso{
  \code{\link{initParallelMode}},
  \code{\link{getConfigHercules}}
}

