\name{getConfigHercules}
\alias{getConfigHercules}
\title{Get the settings of parallelization for hercules}
\usage{
  getConfigHercules()
}
\value{
  the parallelization settings for hercules
}
\description{
  This function returns the current settings for
  parallelization for "hercules", our SGE cluster here at
  DiCITS. This function is intended to be used with
  \code{\link{initParallelMode}}.
}
\seealso{
  \code{\link{initParallelMode}},
  \code{\link{getConfigDefault}}
}

