\name{getFilenameWithoutDatEnding}
\alias{getFilenameWithoutDatEnding}
\title{Remove the string ".dat" or ".DAT" from a filename}
\usage{
  getFilenameWithoutDatEnding(filename)
}
\arguments{
  \item{filename}{a string}
}
\value{
  the \code{filename} string, with ".dat" or ".DAT" removed
  from the end.
}
\description{
  Remove the string ".dat" or ".DAT" from a filename
}

