\name{getNormParams}
\alias{getNormParams}
\alias{getNormParams.aTS}
\title{Get the normalization parameters of a time series}
\usage{
  getNormParams(obj, ...)

  \method{getNormParams}{aTS} (obj, ...)
}
\arguments{
  \item{obj}{the \code{\link{aTS}} object}

  \item{...}{additional function parameters (currently not
  used)}
}
\description{
  This function is a getter function for the
  \code{normParams} atsAttribute.
}

