\name{groupATSListByPartitions}
\alias{groupATSListByPartitions}
\title{Group an \code{\link{ATSList}} by its partitions}
\usage{
  groupATSListByPartitions(atsList)
}
\arguments{
  \item{atsList}{the \code{\link{ATSList}} to which to
  apply the function to}
}
\value{
  a list with two hierarchies
}
\description{
  This method sorts the ATSList according to its
  partitions. It uses the "tsname" attribute of the
  \code{\link{aTS}} objects for that.
}
\details{
  TODO: Where is this used?
}

