\name{linearityTest}
\alias{linearityTest}
\alias{linearityTest.eTS}
\title{LM linearity testing against 2 regime genericStar. Performs a 3rd order Taylor expansion LM test.}
\usage{
  linearityTest(obj, ...)

  \method{linearityTest}{eTS} (obj, rob = FALSE,
    sig = 0.95, trace = TRUE, ...)
}
\arguments{
  \item{obj}{the time series object}

  \item{...}{additional arguments (not used)}

  \item{rob}{TODO: What is this parameter?}

  \item{sig}{the significance level}

  \item{trace}{generate debug output?}
}
\description{
  TODO: This function is from tsDyn
}

