\name{loadKeelDatDirToATSList}
\alias{loadKeelDatDirToATSList}
\title{Load a directory of KEEL dat files into an ATSList}
\usage{
  loadKeelDatDirToATSList(path)
}
\arguments{
  \item{path}{the path of the directory with the dat files}
}
\description{
  Load a directory of KEEL dat files into an ATSList
}

