\name{loadKeelFile}
\alias{loadKeelFile}
\title{Rudimentary parser for KEEL files.}
\usage{
  loadKeelFile(filename, noColumns = 1)
}
\arguments{
  \item{filename}{the name of the KEEL inputfile}

  \item{noColumns}{the number of columns}
}
\description{
  Comments are only allowed in the header
}

