\name{loadMatToATSList}
\alias{loadMatToATSList}
\title{Load mat files into an aTS list}
\usage{
  loadMatToATSList(filename, filenameIndices = NULL)
}
\arguments{
  \item{filename}{the file to read in}

  \item{filenameIndices}{of the positive instances}
}
\value{
  an \code{\link{ATSList}}
}
\description{
  Load mat files into an aTS list
}
\author{
  Mabel
}

