\name{oneLineSummary}
\alias{oneLineSummary}
\title{Generate a "one-line-summary" for an aTS object}
\usage{
  oneLineSummary(aTS, ...)
}
\arguments{
  \item{aTS}{the \code{\link{aTS}} object}

  \item{...}{parameters passed to
  \code{\link{oneLineSummaryAll}}}
}
\description{
  This function calls \code{\link{oneLineSummaryAll}} on
  the \code{coredata} of the \code{\link{aTS}} object.
}

