\name{permutationEntropy}
\alias{permutationEntropy}
\title{calculate the permutation entropy of a given time series}
\usage{
  permutationEntropy(ts, ord = 4)
}
\arguments{
  \item{ts}{the time series to which to apply the function}

  \item{ord}{the order of the permutation entropy}
}
\description{
  Calculates the permutation entropy of the given order for
  the given series.
}

