\name{save.aTS}
\alias{save.aTS}
\title{Save an aTS object to an ".RData" file}
\usage{
  \method{save}{aTS} (aTS, dataPath, name, ...)
}
\arguments{
  \item{aTS}{the \code{\link{aTS}} object of the time
  series}

  \item{dataPath}{the path to which to save the series to}

  \item{name}{the (file-)name to which to save to}

  \item{...}{currently not used}
}
\description{
  This function saves an \code{\link{aTS}} object to an
  ".RData" file.
}

