\name{savePartitions}
\alias{savePartitions}
\title{Save partitioned time series to ATS directory}
\usage{
  savePartitions(partitions, path)
}
\arguments{
  \item{partitions}{a partitioned time series. Usually,
  this is the output of \code{\link{computePartitions}}}

  \item{path}{the directory to which to save the series}
}
\description{
  This function is used to save a partitioned time series
  to an ATS directory.
}

