\name{summary.aTS}
\alias{summary.aTS}
\title{Generic summary function for aTS objects}
\usage{
  \method{summary}{aTS} (object, ...)
}
\arguments{
  \item{object}{the \code{\link{aTS}} object}

  \item{...}{additional function parameters (currently not
  used)}
}
\description{
  This function generates a list with some statistical
  information about the time series in the
  \code{\link{aTS}} object.
}

