\name{toKeelATSDir}
\alias{toKeelATSDir}
\title{Load all files in an ATSDir and save it to a directory of KEEL dat files}
\usage{
  toKeelATSDir(inPath, outPath = NULL)
}
\arguments{
  \item{inPath}{the input path (ATSDir)}

  \item{outPath}{the output path, where to save the dat
  files}
}
\description{
  Load all files in an ATSDir and save it to a directory of
  KEEL dat files
}

