\name{writeATSToKeelFile}
\alias{writeATSToKeelFile}
\title{Save an aTS time series object to a file in KEEL format}
\usage{
  writeATSToKeelFile(aTS, filename, useColNames = FALSE)
}
\arguments{
  \item{aTS}{the aTS object to save}

  \item{filename}{complete path to the file to write}

  \item{useColNames}{should colnames present in the aTS
  object be used for attribute naming?}
}
\description{
  Save an aTS time series object to a file in KEEL format
}

