\name{writeKeelXmlDescription}
\alias{writeKeelXmlDescription}
\title{Write out an XML description for use with KEEL}
\usage{
  writeKeelXmlDescription(atsList, path, filename)
}
\arguments{
  \item{atsList}{the atsList to use in KEEL}

  \item{path}{the directory where to save the xml file}

  \item{filename}{the filename of the xml file}
}
\description{
  Write out an XML description for use with KEEL
}

